Use gopkg.in/yaml.v2 instead
ref: https://github.com/k1LoW/runn/issues/103

+++ operator_test.go
@@ -45,13 +45,13 @@ func TestExpand(t *testing.T) {
 			[]map[string]interface{}{},
 			map[string]interface{}{"one": 1},
 			map[string]string{"key": "{{ vars.one }}"},
-			map[string]interface{}{"key": uint64(1)},
+			map[string]interface{}{"key": 1},
 		},
 		{
 			[]map[string]interface{}{},
 			map[string]interface{}{"one": 1},
 			map[string]string{"key": "{{ vars.one + 1 }}"},
-			map[string]interface{}{"key": uint64(2)},
+			map[string]interface{}{"key": 2},
 		},
 		{
 			[]map[string]interface{}{},
@@ -81,19 +81,19 @@ func TestExpand(t *testing.T) {
 			[]map[string]interface{}{},
 			map[string]interface{}{"map": map[string]interface{}{"foo": "test", "bar": 1}},
 			map[string]string{"map": "{{ vars.map }}"},
-			map[string]interface{}{"map": map[string]interface{}{"foo": "test", "bar": uint64(1)}},
+			map[string]interface{}{"map": map[string]interface{}{"foo": "test", "bar": 1}},
 		},
 		{
 			[]map[string]interface{}{},
 			map[string]interface{}{"array": []interface{}{map[string]interface{}{"foo": "test1", "bar": 1}, map[string]interface{}{"foo": "test2", "bar": 2}}},
 			map[string]string{"array": "{{ vars.array }}"},
-			map[string]interface{}{"array": []interface{}{map[string]interface{}{"foo": "test1", "bar": uint64(1)}, map[string]interface{}{"foo": "test2", "bar": uint64(2)}}},
+			map[string]interface{}{"array": []interface{}{map[string]interface{}{"foo": "test1", "bar": 1}, map[string]interface{}{"foo": "test2", "bar": 2}}},
 		},
 		{
 			[]map[string]interface{}{},
 			map[string]interface{}{"float": float64(1)},
 			map[string]string{"float": "{{ vars.float }}"},
-			map[string]interface{}{"float": uint64(1)},
+			map[string]interface{}{"float": 1},
 		},
 		{
 			[]map[string]interface{}{},
@@ -105,7 +105,7 @@ func TestExpand(t *testing.T) {
 			[]map[string]interface{}{},
 			map[string]interface{}{"float": float64(1.00)},
 			map[string]string{"float": "{{ vars.float }}"},
-			map[string]interface{}{"float": uint64(1)},
+			map[string]interface{}{"float": 1},
 		},
 		{
 			[]map[string]interface{}{},
@@ -123,7 +123,7 @@ func TestExpand(t *testing.T) {
 			[]map[string]interface{}{},
 			map[string]interface{}{"uint64": uint64(4600)},
 			map[string]string{"uint64": "{{ vars.uint64 }}"},
-			map[string]interface{}{"uint64": uint64(4600)},
+			map[string]interface{}{"uint64": 4600},
 		},
 	}
 	for _, tt := range tests {
